# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey Service OpenApi3 Documentation",
        "description": "Behandam Survey Service Api documentation",
        "contact": {
            "email": "admin@kermany.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "LocalHost API Server"
        },
        {
            "url": "https://debug.behaminplus.ir/survey-service",
            "description": "Debug API Server"
        },
        {
            "url": "https://develop.behaminplus.ir/survey-service",
            "description": "DEVELOP API Server"
        }
    ],
    "paths": {
...
        "/api/v1/packages/types": {
            "get": {
                "tags": [
                    "Package Types"
                ],
                "summary": "get list of all package types : if empty you should run seeder",
                "operationId": "getListOfPackageTypes",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/packages/{packageId}/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "get reports by package id",
                "operationId": "getReportByPackageId",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "161"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "access-token": {
                "type": "http",
                "description": "Use Custom access token",
                "name": "AccessToken",
                "scheme": "access-token"
            }
        }
    },
...
    "security": [
        []
    ]
}